Add aggregator CI pipeline
Add workflow for linting, typechecking, and running tests for aggregator.
Break up aggregator premerge program into additional typescript check & todo linting programs.
Update default aggregator private keys in .env.example to hardhat node values.
Add fundDeployer hardhat task to simplify local & CI contract deployment.
Remove steps from local development doc.
Default contracts & clients setup action to use NodeJS 16.x.
