+++ .github/workflows/aggregator-ci.yml
@@ -1,0 +1,78 @@
+name: aggregator
+
+on:
+  push:
+    branches:
+      - 'main'
+    paths:
+      - 'aggregator/**'
+  pull_request:
+    branches:
+      - 'main'
+    paths:
+      - 'aggregator/**'
+
+defaults:
+  run:
+    working-directory: ./aggregator
+
+env:
+  DENO_VERSION: 1.x
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - uses: denoland/setup-deno@v1
+      with:
+        deno-version: ${{ env.DENO_VERSION }}
+    - run: deno lint .
+
+  todos-fixmes:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - uses: denoland/setup-deno@v1
+      with:
+        deno-version: ${{ env.DENO_VERSION }}
+    - run: ./programs/lintTodos.ts
+
+  typescript:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - uses: denoland/setup-deno@v1
+      with:
+        deno-version: ${{ env.DENO_VERSION }}
+    - run: ./programs/checkTs.ts
+
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - uses: denoland/setup-deno@v1
+      with:
+        deno-version: ${{ env.DENO_VERSION }}
+    - uses: ./.github/actions/setup-contracts-clients
+
+    # Setup contracts
+    - working-directory: ./contracts
+      run: yarn hardhat node &
+    - working-directory: ./contracts
+      run: yarn hardhat fundDeployer --network gethDev
+    - working-directory: ./contracts
+      run: yarn hardhat run scripts/deploy_all.ts --network gethDev
+
+    - working-directory: ./
+      run: docker-compose up -d postgres
+    - run: cp .env.example .env
+    - run: deno test --allow-net --allow-env --allow-read --unstable
+
+    # Cleanup
+    - working-directory: ./
+      run: docker-compose down
