+++ .github/workflows/contracts-ci.yml
@@ -1,4 +1,4 @@
-name: Contracts
+name: contracts
 
 on:
   push:
@@ -18,26 +18,19 @@ defaults:
   run:
     working-directory: ./contracts
 
-env:
-  NODEJS_VERSION: 16.x
-
 jobs:
-  contracts-lint:
+  lint:
     runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v3
     - uses: ./.github/actions/setup-contracts-clients
-      with:
-        node-version: ${{ env.NODEJS_VERSION }}
     - run: yarn lint
 
-  contracts-test:
+  test:
     runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v3
     - uses: ./.github/actions/setup-contracts-clients
-      with:
-        node-version: ${{ env.NODEJS_VERSION }}
     - run: yarn test
