+++ aggregator/.env.example
@@ -6,8 +6,8 @@ ORIGIN=http://localhost:3000
 PORT=3000
 
 NETWORK_CONFIG_PATH=../contracts/networks/local.json
-PRIVATE_KEY_AGG=0x0000000000000000000000000000000000000000000000000000000000000a99
-PRIVATE_KEY_ADMIN=
+PRIVATE_KEY_AGG=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
+PRIVATE_KEY_ADMIN=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d
 TEST_BLS_WALLETS_SECRET=test-bls-wallets-secret
 
 PG_HOST=localhost
