+++ aggregator/programs/checkTs.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env -S deno run --unstable --allow-run --allow-read --allow-write --allow-env
+
+import { checkTs } from "./helpers/typescript.ts";
+
+await checkTs();
