+++ contracts/hardhat.config.ts
@@ -8,18 +8,34 @@ import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import "hardhat-gas-reporter";
 import "solidity-coverage";
+import defaultDeployerWallets from "./shared/helpers/defaultDeployerWallet";
 
 dotenv.config();
 
 // This is a sample Hardhat task. To learn how to create your own go to
 // https://hardhat.org/guides/create-task.html
-task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
+task("accounts", "Prints the list of accounts", async (_taskArgs, hre) => {
   const accounts = await hre.ethers.getSigners();
 
   for (const account of accounts) {
     console.log(account.address);
   }
 });
+
+task("fundDeployer", "Sends ETH to create2Deployer contract from first signer")
+  .addOptionalParam("amount", "Amount of ETH to send", "1.0")
+  .setAction(async ({ amount }: { amount: string }, hre) => {
+    const [account0] = await hre.ethers.getSigners();
+    const deployerAddress = defaultDeployerWallets(hre.ethers).address;
+
+    console.log(`${account0.address} -> ${deployerAddress} ${amount} ETH`);
+
+    const txnRes = await account0.sendTransaction({
+      to: deployerAddress,
+      value: hre.ethers.utils.parseEther(amount),
+    });
+    await txnRes.wait();
+  });
 
 // Do any needed pre-test setup here.
 task("test").setAction(async (_taskArgs, _hre, runSuper) => {
