+++ docs/remote_development.md
@@ -2,13 +2,13 @@
 
 These steps will setup this repo on your machine for targeting a remote chain, such as an EVM compatible L2.
 
-Follow the instructions for [Local Development](./local_development.md), replacing the section titled `Chain & Contracts` with the steps below.
+Follow the instructions for [Local Development](./local_development.md), replacing the sections titled `Chain` and `Contracts` with the steps below.
 
 ## Deploy Contracts
 
 ### Deployer account
 
-BLS Wallet contract deploys use `CREATE2` to maintain consistent addresses across networks. As such, a create2 deployer account is used and listed in `./contracts/.env` under the environment variables `DEPLOYER_MNEMONIC` & `DEPLOYER_SET_INDEX`. That account will need to be funded in order to deploy the contracts.
+BLS Wallet contract deploys use `CREATE2` to maintain consistent addresses across networks. As such, a create2 deployer contract is used and listed in `./contracts/.env` under the environment variables `DEPLOYER_MNEMONIC` & `DEPLOYER_SET_INDEX`. The HD address will need to be funded in order to deploy the contracts.
 
 If you do not need consistent addresses, for example on a local or testnet network, you can replace the `DEPLOYER_MNEMONIC` with another seed phrase which already has a funded account.
 
@@ -26,7 +26,7 @@ yarn hardhat run scripts/0_deploy_precompile_cost_estimator.ts --network YOUR_NE
 ```
 Copy the address that is output.
 
-Update `./contracts/contracts/lib/hubble-contracts/contracts/libs/BLS.sol`'s `COST_ESTIMATOR_ADDRESS` to the value of that address;
+Update `./contracts/contracts/lib/hubble-contracts/contracts/libs/BLS.sol`'s `COST_ESTIMATOR_ADDRESS` to the value of that address if it is different:
 ```solidity
 ...
 address private constant COST_ESTIMATOR_ADDRESS = YOUR_NETWORKS_PRECOMPILE_COST_ESTIMATOR_ADDRESS;
